+++ alacritty_terminal/CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## 0.26.1-dev
 
+### Added
+
+- OSC 7 working-directory URI tracking through `Term::current_directory`
+
 ### Fixed
 
 - Panic when the PTY could not be set to non-blocking